requirements.txt
typer
pyzipper
requests
rich
lief
lxml
pyyaml
yara-python
paramiko
python-dotenv
python-crontab
psutil